+++ oci/kms_key_resource_test.go
@@ -3,87 +3,43 @@
 package oci
 
 import (
-	"context"
 	"fmt"
-	"strconv"
 	"testing"
 	"time"
 
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
-	"github.com/oracle/oci-go-sdk/common"
-
-	oci_kms "github.com/oracle/oci-go-sdk/keymanagement"
 
 	"github.com/terraform-providers/terraform-provider-oci/httpreplay"
 )
 
 var (
-	KeyRequiredOnlyResource = KeyResourceDependencies +
-		generateResourceFromRepresentationMap("oci_kms_key", "test_key", Required, Create, keyRepresentation)
+	KeyResourceVirtualDependencyConfig = KeyResourceVirtualDependencies + DefinedTagsDependencies +
+		generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Create, keyVirtualRepresentation)
 
-	KeyResourceConfig = KeyResourceDependencies +
-		generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Update, keyRepresentation)
+	keyVirtualRepresentation = getMultipleUpdatedRepresenationCopy([]string{"management_endpoint", "time_of_deletion"}, []interface{}{
+		Representation{repType: Required, create: `${data.oci_kms_vault.test_virtual_vault.management_endpoint}`},
+		Representation{repType: Required, create: deletionTime.Format(time.RFC3339Nano)}}, keyRepresentation)
+	keyVirtualDataSourceRepresentation         = getUpdatedRepresentationCopy("management_endpoint", Representation{repType: Required, create: `${data.oci_kms_vault.test_virtual_vault.management_endpoint}`}, keyDataSourceRepresentation)
+	keyVirtualSingularDataSourceRepresentation = getUpdatedRepresentationCopy("management_endpoint", Representation{repType: Required, create: `${data.oci_kms_vault.test_virtual_vault.management_endpoint}`}, keySingularDataSourceRepresentation)
 
-	keySingularDataSourceRepresentation = map[string]interface{}{
-		"key_id":              Representation{repType: Required, create: `${oci_kms_key.test_key.id}`},
-		"management_endpoint": Representation{repType: Required, create: `${data.oci_kms_vault.test_vault.management_endpoint}`},
-	}
-
-	keyDataSourceRepresentation = map[string]interface{}{
-		"compartment_id":      Representation{repType: Required, create: `${var.compartment_id}`},
-		"management_endpoint": Representation{repType: Required, create: `${data.oci_kms_vault.test_vault.management_endpoint}`},
-		"filter":              RepresentationGroup{Required, keyDataSourceFilterRepresentation}}
-	keyDataSourceFilterRepresentation = map[string]interface{}{
-		"name":   Representation{repType: Required, create: `id`},
-		"values": Representation{repType: Required, create: []string{`${oci_kms_key.test_key.id}`}},
-	}
-
-	deletionTime = time.Now().UTC().AddDate(0, 0, 8).Truncate(time.Millisecond)
-
-	keyRepresentation = map[string]interface{}{
-		"compartment_id":      Representation{repType: Required, create: `${var.compartment_id}`},
-		"display_name":        Representation{repType: Required, create: `Key C`, update: `displayName2`},
-		"key_shape":           RepresentationGroup{Required, keyKeyShapeRepresentation},
-		"management_endpoint": Representation{repType: Required, create: `${data.oci_kms_vault.test_vault.management_endpoint}`},
-		"desired_state":       Representation{repType: Optional, create: `ENABLED`, update: `DISABLED`},
-		"defined_tags":        Representation{repType: Optional, create: `${map("${oci_identity_tag_namespace.tag-namespace1.name}.${oci_identity_tag.tag1.name}", "value")}`, update: `${map("${oci_identity_tag_namespace.tag-namespace1.name}.${oci_identity_tag.tag1.name}", "updatedValue")}`},
-		"freeform_tags":       Representation{repType: Optional, create: map[string]string{"Department": "Finance"}, update: map[string]string{"Department": "Accounting"}},
-		"time_of_deletion":    Representation{repType: Optional, create: deletionTime.Format(time.RFC3339Nano)},
-	}
-	keyKeyShapeRepresentation = map[string]interface{}{
-		"algorithm": Representation{repType: Required, create: `AES`},
-		"length":    Representation{repType: Required, create: `16`},
-	}
-	KeyResourceDependencies = `
-	variable "vault_ids" {
+	KeyResourceVirtualDependencies = `
+	variable "virtual_vault_ids" {
 		type = "map"
 		default = {
-			us-phoenix-1 = "ocid1.vault.oc1.phx.avnzdivwaadfa.abyhqljrmb7herjt4gz64avusyue25grswwsqc5x75im7vtg4x7yfgszqkfa"
-			us-ashburn-1 = "ocid1.vault.oc1.iad.annnb3f4aacuu.abuwcljrumuxamzquswnwvgvqdyc76v4e6lyo4372wcjvtdxrhxdc6qxlupq"
+			us-phoenix-1 = "ocid1.vault.oc1.phx.a5ov6eneaafna.abyhqljs563nxjpbh7hrx73ivi6mjc2xg3q7ljtxmaczys6qoxwg3vyordxq"
+			us-ashburn-1 = "ocid1.vault.oc1.iad.bbo7h3seaaeug.abuwcljskeh5tco23lpk5hkijjrlrj64q5afzz3qbe25ku3b4n5ozn66qr5a"	
 		}
 	}
-	data "oci_kms_vault" "test_vault" {
+	data "oci_kms_vault" "test_virtual_vault" {
 		#Required
-		vault_id = "${var.vault_ids[var.region]}"
-	}
-	`
-	KeyResourceDependencyConfig = KeyResourceDependencies + `
-	data "oci_kms_keys" "test_keys_dependency" {
-		#Required
-		compartment_id = "${var.tenancy_ocid}"
-		management_endpoint = "${data.oci_kms_vault.test_vault.management_endpoint}"
-
-		filter {
-    		name = "state"
-    		values = ["ENABLED"]
-        }
+		vault_id = "${var.virtual_vault_ids[var.region]}"
 	}
 	`
 )
 
-func TestKmsKeyResource_basic(t *testing.T) {
-	httpreplay.SetScenario("TestKmsKeyResource_basic")
+func TestResourceKmsKeyResource_basic(t *testing.T) {
+	httpreplay.SetScenario("TestResourceKmsKeyResource_basic")
 	defer httpreplay.SaveScenario()
 
 	provider := testAccProvider
@@ -110,8 +66,8 @@ func TestKmsKeyResource_basic(t *testing.T) {
 		Steps: []resource.TestStep{
 			// verify create
 			{
-				Config: config + compartmentIdVariableStr + KeyResourceDependencies + DefinedTagsDependencies +
-					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Required, Create, keyRepresentation),
+				Config: config + compartmentIdVariableStr + KeyResourceVirtualDependencies + DefinedTagsDependencies +
+					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Required, Create, keyVirtualRepresentation),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "compartment_id", compartmentId),
 					resource.TestCheckResourceAttr(resourceName, "display_name", "Key C"),
@@ -132,8 +88,8 @@ func TestKmsKeyResource_basic(t *testing.T) {
 			},
 			// verify create with optionals
 			{
-				Config: config + compartmentIdVariableStr + KeyResourceDependencies + DefinedTagsDependencies +
-					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Create, keyRepresentation),
+				Config: config + compartmentIdVariableStr + KeyResourceVirtualDependencies + DefinedTagsDependencies +
+					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Create, keyVirtualRepresentation),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "compartment_id", compartmentId),
 					resource.TestCheckResourceAttrSet(resourceName, "current_key_version"),
@@ -151,11 +107,6 @@ func TestKmsKeyResource_basic(t *testing.T) {
 
 					func(s *terraform.State) (err error) {
 						resId, err = fromInstanceState(s, resourceName, "id")
-						if isEnableExportCompartment, _ := strconv.ParseBool(getEnvSettingWithDefault("enable_export_compartment", "false")); isEnableExportCompartment {
-							if errExport := testExportCompartmentWithResourceName(&resId, &compartmentId, resourceName); errExport != nil {
-								return errExport
-							}
-						}
 						return err
 					},
 				),
@@ -163,9 +114,9 @@ func TestKmsKeyResource_basic(t *testing.T) {
 
 			// verify update to the compartment (the compartment will be switched back in the next step)
 			{
-				Config: config + compartmentIdVariableStr + compartmentIdUVariableStr + KeyResourceDependencies + DefinedTagsDependencies +
+				Config: config + compartmentIdVariableStr + compartmentIdUVariableStr + KeyResourceVirtualDependencies + DefinedTagsDependencies +
 					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Create,
-						representationCopyWithNewProperties(keyRepresentation, map[string]interface{}{
+						representationCopyWithNewProperties(keyVirtualRepresentation, map[string]interface{}{
 							"compartment_id": Representation{repType: Required, create: `${var.compartment_id_for_update}`},
 						})),
 				Check: resource.ComposeAggregateTestCheckFunc(
@@ -194,8 +145,8 @@ func TestKmsKeyResource_basic(t *testing.T) {
 
 			// verify updates to updatable parameters
 			{
-				Config: config + compartmentIdVariableStr + KeyResourceDependencies + DefinedTagsDependencies +
-					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Update, keyRepresentation),
+				Config: config + compartmentIdVariableStr + KeyResourceVirtualDependencies + DefinedTagsDependencies +
+					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Update, keyVirtualRepresentation),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "compartment_id", compartmentId),
 					resource.TestCheckResourceAttrSet(resourceName, "current_key_version"),
@@ -222,9 +173,9 @@ func TestKmsKeyResource_basic(t *testing.T) {
 			// verify datasource
 			{
 				Config: config +
-					generateDataSourceFromRepresentationMap("oci_kms_keys", "test_keys", Optional, Update, keyDataSourceRepresentation) +
-					compartmentIdVariableStr + KeyResourceDependencies + DefinedTagsDependencies +
-					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Update, keyRepresentation),
+					generateDataSourceFromRepresentationMap("oci_kms_keys", "test_keys", Optional, Update, keyVirtualDataSourceRepresentation) +
+					compartmentIdVariableStr + KeyResourceVirtualDependencies + DefinedTagsDependencies +
+					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Update, keyVirtualRepresentation),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr(datasourceName, "compartment_id", compartmentId),
 
@@ -242,8 +193,9 @@ func TestKmsKeyResource_basic(t *testing.T) {
 			// verify singular datasource
 			{
 				Config: config +
-					generateDataSourceFromRepresentationMap("oci_kms_key", "test_key", Required, Create, keySingularDataSourceRepresentation) +
-					compartmentIdVariableStr + KeyResourceConfig + DefinedTagsDependencies,
+					generateDataSourceFromRepresentationMap("oci_kms_key", "test_key", Required, Create, keyVirtualSingularDataSourceRepresentation) +
+					compartmentIdVariableStr + KeyResourceVirtualDependencies +
+					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Update, keyVirtualRepresentation) + DefinedTagsDependencies,
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttrSet(singularDatasourceName, "key_id"),
 
@@ -263,12 +215,13 @@ func TestKmsKeyResource_basic(t *testing.T) {
 			},
 			// remove singular datasource from previous step so that it doesn't conflict with import tests
 			{
-				Config: config + compartmentIdVariableStr + KeyResourceConfig + DefinedTagsDependencies,
+				Config: config + compartmentIdVariableStr + KeyResourceVirtualDependencies +
+					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Update, keyVirtualRepresentation) + DefinedTagsDependencies,
 			},
 			// revert the updates
 			{
-				Config: config + compartmentIdVariableStr + KeyResourceDependencies + DefinedTagsDependencies +
-					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Create, keyRepresentation),
+				Config: config + compartmentIdVariableStr + KeyResourceVirtualDependencies + DefinedTagsDependencies +
+					generateResourceFromRepresentationMap("oci_kms_key", "test_key", Optional, Create, keyVirtualRepresentation),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "display_name", "Key C"),
 					resource.TestCheckResourceAttr(resourceName, "state", "ENABLED"),
@@ -295,63 +248,4 @@ func TestKmsKeyResource_basic(t *testing.T) {
 			},
 		},
 	})
-}
-
-func testAccCheckKMSKeyDestroy(s *terraform.State) error {
-	noResourceFound := true
-	for _, rs := range s.RootModule().Resources {
-		if rs.Type == "oci_kms_key" {
-			client, err := testAccProvider.Meta().(*OracleClients).KmsManagementClient(rs.Primary.Attributes["management_endpoint"])
-			if err != nil {
-				return err
-			}
-
-			noResourceFound = false
-			request := oci_kms.GetKeyRequest{}
-
-			tmp := rs.Primary.ID
-			request.KeyId = &tmp
-
-			request.RequestMetadata.RetryPolicy = getRetryPolicy(true, "kms")
-
-			response, err := client.GetKey(context.Background(), request)
-
-			if err == nil {
-				deletedLifecycleStates := map[string]bool{
-					string(oci_kms.KeyLifecycleStateSchedulingDeletion): true,
-					string(oci_kms.KeyLifecycleStatePendingDeletion):    true,
-				}
-				if _, ok := deletedLifecycleStates[string(response.LifecycleState)]; !ok {
-					//resource lifecycle state is not in expected deleted lifecycle states.
-					return fmt.Errorf("resource lifecycle state: %s is not in expected deleted lifecycle states", response.LifecycleState)
-				}
-
-				if !response.TimeOfDeletion.Equal(deletionTime) && !httpreplay.ModeRecordReplay() {
-					return fmt.Errorf("resource time_of_deletion: %s is not set to %s", response.TimeOfDeletion.Format(time.RFC3339Nano), deletionTime.Format(time.RFC3339Nano))
-				}
-				//resource lifecycle state is in expected deleted lifecycle states. continue with next one.
-				continue
-			}
-
-			//Verify that exception is for '404 not found'.
-			if failure, isServiceError := common.IsServiceError(err); !isServiceError || failure.GetHTTPStatusCode() != 404 {
-				return err
-			}
-		}
-	}
-	if noResourceFound {
-		return fmt.Errorf("at least one resource was expected from the state file, but could not be found")
-	}
-
-	return nil
-}
-
-func keyImportId(state *terraform.State) (string, error) {
-	for _, rs := range state.RootModule().Resources {
-		if rs.Type == "oci_kms_key" {
-			return fmt.Sprintf("managementEndpoint/%s/keys/%s", rs.Primary.Attributes["management_endpoint"], rs.Primary.ID), nil
-		}
-	}
-
-	return "", fmt.Errorf("unable to create import id as no resource of type oci_kms_key in state")
 }